removed bug
